requirements.txt
ipython[notebook]
numpy
scipy
matplotlib
seaborn
tweepy
python-decouple
